chore: add minor CR improvements
